% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_patterns.R
\name{all.patterns}
\alias{all.patterns}
\title{Generate All Response Patterns}
\usage{
\method{all}{patterns}(J, pat)
}
\arguments{
\item{J}{Number of items}

\item{pat}{A vector for the item response categories}
}
\value{
A matrix of all possible response patterns (each row is one pattern)
}
\description{
This function generates all possible response patterns for `J` items with `levels`.
}
\keyword{internal}
