% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateLcm.R
\name{estimateLcm}
\alias{estimateLcm}
\title{LSCAT Calibration}
\usage{
estimateLcm(X, ic = "aic", method = "all", saveLCMparams = FALSE)
}
\arguments{
\item{X}{A data frame with the available responses on all items}

\item{ic}{Information criterion; One of "aic", "aic3", "aicc", "bic". Default is "aic".}

\item{method}{Default is "all" (all possible responses).
Because the number of rows may be too large, several row-reduction methods have been added: "mod" (model-based responses), "emp" (empirical responses) and "inf" (most informative item responses).}

\item{saveLCMparams}{Logical. Default is `FALSE`. If `TRUE` the function returns also the class weights (W), and
the conditional response probabilities (P) under the best fitting Latent Class Model.}
}
\value{
An object of class list which includes
X : the data.frame used to calibrate the LCM.
R : A matrix with the item-score vectors used.
RP: A vector containing the probability of each item-score vector
nCat: The number of item categories
K: The number of classes in the best fitting LCM
}
\description{
This function first fits a Latent Class Model (LCM) using `poLCA::` and keeps the best fitting model according to a specified Information Criterion.
Then, computes the item-score density (π) under one of 4 choices: "mod", "emp", "inf" and "all"
}
\references{
Psychogyiopoulos, A., Smits, N., Van der Ark, L.A.: Estimating the joint item- score density using an unrestricted latent class model: Advancing flexibility in computerized adaptive testing. Journal of Computerized Adaptive Testing (in press)
Van der Ark, L.A., Smits, N.: Computerized adaptive testing without IRT for flexible measurement and prediction. In: Essays on Contemporary Psychometrics, pp. 369–388. Springer, (2023). https://doi.org/10.1007/978-3-031-10370-4 19
}
